part 4

+++ part4/controllers/blogs.js
@@ -1,37 +1,15 @@
 const blogsRouter = require('express').Router()
 const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const User = require('../models/user');
+const blog = require('../models/blog');
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({});
-    response.json(blogs);
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+
+  response.json(blogs)
 });
-
-  
-blogsRouter.post('/', async (request, response) => {
-    const body = request.body
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-
-    if(!decodedToken.id){
-      return response.status(401).json({ error: 'token invalid' })
-    }
-    const user = await User.findById(body.userId)
-
-    const blog = new Blog({
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: body.likes ? body.likes : 0,
-      user: user.id
-    })
-
-    const savedBlog = await blog.save()
-    user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
-    
-    response.status(201).json(savedBlog)
-})
 
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
@@ -41,10 +19,48 @@ blogsRouter.get('/:id', async (request, response) => {
     response.status(404).end()
   }
 })
+  
+blogsRouter.post('/', async (request, response) => {
+    const body = request.body
+    const user = request.user
+
+    if(!user) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes ? body.likes : 0,
+      user: user.id
+    })
+
+    if(body.title === undefined || body.url === undefined){
+      response.status(400).end()
+    } else {
+      const savedBlog = await blog.save()
+      user.blogs = user.blogs.concat(savedBlog._id)
+      await user.save()
+      
+      response.status(201).json(savedBlog)
+    }
+})
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const user = request.user
+
+  if(!user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if(blog.user.toString() === request.user.id){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'unauthorized to delete the blog' })
+  }
 })
 
 blogsRouter.put('/:id', async (request,response) => {
